# private/protocol/json: cache SDKShapeTraits tag