More work on https://github.com/aws/aws-sdk-go/issues/377

The `reflect.FieldByName` call is expensive. This caches that lookup.

Benchmark (with SDKShapeTraits field) to capture the cost of this call:

```
benchmark                 old ns/op     new ns/op     delta
BenchmarkBuildJSON-4      15180         9690          -36.17%

benchmark                 old allocs     new allocs     delta
BenchmarkBuildJSON-4      98             62             -36.73%

benchmark                 old bytes     new bytes     delta
BenchmarkBuildJSON-4      3136          1408          -55.10%
```

A downside of using a global cache like this is that it will stay around taking up memory even if the process will never (un)marshal that type again. At least in this case the cache is bounded to types with a `SDKShapeTraits` field.

Another (more invasive?) solution would be to have the (un)marshal handlers own the cache so it can be garbage collected when a client is no longer referenced. But this could lead to a duplication of the cache if there are multiple client instances for a service.