@bpot 2015-11-13
Deltas for `make bench-protocol` against master+the `build_test.go` change:

```
benchmark                                                                      old ns/op     new ns/op     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-4               51114         50054         -2.07%
BenchmarkEC2QueryBuild_Simple_ec2AttachNetworkInterface-4                      12863         13034         +1.33%
BenchmarkBuildJSON-4                                                           14813         9617          -35.08%
BenchmarkStdlibJSON-4                                                          7061          7014          -0.67%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-4                                 18704         14767         -21.05%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-4                                18671         14005         -24.99%
BenchmarkEncodingJSONMarshal_Simple_dynamodbPutItem-4                          6377          6395          +0.28%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-4               273837        215841        -21.18%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-4                   8867          8984          +1.32%
BenchmarkEncodingJSONMarshal_Complex_elastictranscoderCreateJobInput-4         37970         39407         +3.78%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-4            14139         13099         -7.36%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-4                10115         10440         +3.21%
BenchmarkEncodingJSONMarshal_Simple_elastictranscoderListJobsByPipeline-4      1096          1089          -0.64%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-4                   405664        405967        +0.07%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-4           18864         18716         -0.78%
BenchmarkEncodingXMLMarshal_Simple_cloudfrontDeleteStreamingDistribution-4     5132          5217          +1.66%

benchmark                                                                      old allocs     new allocs     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-4               237            237            +0.00%
BenchmarkEC2QueryBuild_Simple_ec2AttachNetworkInterface-4                      67             67             +0.00%
BenchmarkBuildJSON-4                                                           98             62             -36.73%
BenchmarkStdlibJSON-4                                                          13             13             +0.00%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-4                                 105            78             -25.71%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-4                                103            76             -26.21%
BenchmarkEncodingJSONMarshal_Simple_dynamodbPutItem-4                          8              8              +0.00%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-4               1131           798            -29.44%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-4                   59             59             +0.00%
BenchmarkEncodingJSONMarshal_Complex_elastictranscoderCreateJobInput-4         7              7              +0.00%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-4            80             71             -11.25%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-4                57             57             +0.00%
BenchmarkEncodingJSONMarshal_Simple_elastictranscoderListJobsByPipeline-4      2              2              +0.00%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-4                   1832           1832           +0.00%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-4           88             88             +0.00%
BenchmarkEncodingXMLMarshal_Simple_cloudfrontDeleteStreamingDistribution-4     9              9              +0.00%

benchmark                                                                      old bytes     new bytes     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-4               10922         10922         +0.00%
BenchmarkEC2QueryBuild_Simple_ec2AttachNetworkInterface-4                      3696          3696          +0.00%
BenchmarkBuildJSON-4                                                           3136          1408          -55.10%
BenchmarkStdlibJSON-4                                                          1200          1200          +0.00%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-4                                 4289          2992          -30.24%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-4                                4225          2928          -30.70%
BenchmarkEncodingJSONMarshal_Simple_dynamodbPutItem-4                          808           808           +0.00%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-4               41276         25287         -38.74%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-4                   2800          2800          +0.00%
BenchmarkEncodingJSONMarshal_Complex_elastictranscoderCreateJobInput-4         4905          4905          +0.00%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-4            4368          3936          -9.89%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-4                3328          3328          +0.00%
BenchmarkEncodingJSONMarshal_Simple_elastictranscoderListJobsByPipeline-4      120           120           +0.00%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-4                   103354        103383        +0.03%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-4           9104          9104          +0.00%
BenchmarkEncodingXMLMarshal_Simple_cloudfrontDeleteStreamingDistribution-4     4656          4656          +0.00%
```