@jasdel 2015-11-14
Thanks again for the PR @bpot, We're definitely whittling down the allocations of the marshalers. Using a global cache seems a little heavy handed for addressing the lookup of these types.  I'm curious if we could instead use interfaces to extract this information instead of performing lookup via reflect.

Something like moving the `payload:"field"` tag from SDKShapeTraits and move the value into a method on the type. It would be great to remove the `metadata<T>` and `SDKShapeTraits` fields, but that is a secondary.

I pushed [marshalPayloadInf](https://github.com/aws/aws-sdk-go/tree/marshalPayloadInf) branch that we can take a look at. Do you see similar performance differences?

The follow benchmark compares caching vs using a payload interface.
```
benchmark                                                                      old ns/op     new ns/op     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-8               48102         46714         -2.89%
BenchmarkEC2QueryBuild_Simple_ec2AttachNetworkInterface-8                      13399         13266         -0.99%
BenchmarkBuildJSON-8                                                           10548         9839          -6.72%
BenchmarkStdlibJSON-8                                                          8374          7651          -8.63%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-8                                 13871         13398         -3.41%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-8                                13705         13062         -4.69%
BenchmarkEncodingJSONMarshal_Simple_dynamodbPutItem-8                          6103          6064          -0.64%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-8               217430        214609        -1.30%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-8                   8455          7145          -15.49%
BenchmarkEncodingJSONMarshal_Complex_elastictranscoderCreateJobInput-8         37409         37262         -0.39%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-8            13425         10628         -20.83%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-8                10420         8807          -15.48%
BenchmarkEncodingJSONMarshal_Simple_elastictranscoderListJobsByPipeline-8      1052          955           -9.22%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-8                   370145        377077        +1.87%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-8           17148         14620         -14.74%
BenchmarkEncodingXMLMarshal_Simple_cloudfrontDeleteStreamingDistribution-8     3682          3563          -3.23%

benchmark                                                                      old allocs     new allocs     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-8               237            237            +0.00%
BenchmarkEC2QueryBuild_Simple_ec2AttachNetworkInterface-8                      67             67             +0.00%
BenchmarkBuildJSON-8                                                           49             49             +0.00%
BenchmarkStdlibJSON-8                                                          13             13             +0.00%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-8                                 52             52             +0.00%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-8                                50             50             +0.00%
BenchmarkEncodingJSONMarshal_Simple_dynamodbPutItem-8                          8              8              +0.00%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-8               798            781            -2.13%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-8                   59             51             -13.56%
BenchmarkEncodingJSONMarshal_Complex_elastictranscoderCreateJobInput-8         7              7              +0.00%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-8            71             54             -23.94%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-8                57             49             -14.04%
BenchmarkEncodingJSONMarshal_Simple_elastictranscoderListJobsByPipeline-8      2              2              +0.00%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-8                   1832           1815           -0.93%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-8           88             71             -19.32%
BenchmarkEncodingXMLMarshal_Simple_cloudfrontDeleteStreamingDistribution-8     9              9              +0.00%

benchmark                                                                      old bytes     new bytes     delta
BenchmarkEC2QueryBuild_Complex_ec2AuthorizeSecurityGroupEgress-8               10921         10922         +0.01%
BenchmarkEC2QueryBuild_Simple_ec2AttachNetworkInterface-8                      3696          3696          +0.00%
BenchmarkBuildJSON-8                                                           1200          1200          +0.00%
BenchmarkStdlibJSON-8                                                          1200          1200          +0.00%
BenchmarkJSONRPCBuild_Simple_dynamodbPutItem-8                                 2577          2577          +0.00%
BenchmarkJSONUtilBuild_Simple_dynamodbPutItem-8                                2513          2513          +0.00%
BenchmarkEncodingJSONMarshal_Simple_dynamodbPutItem-8                          808           808           +0.00%
BenchmarkRESTJSONBuild_Complex_elastictranscoderCreateJobInput-8               25293         24524         -3.04%
BenchmarkRESTBuild_Complex_elastictranscoderCreateJobInput-8                   2800          2464          -12.00%
BenchmarkEncodingJSONMarshal_Complex_elastictranscoderCreateJobInput-8         4906          4906          +0.00%
BenchmarkRESTJSONBuild_Simple_elastictranscoderListJobsByPipeline-8            3936          3104          -21.14%
BenchmarkRESTBuild_Simple_elastictranscoderListJobsByPipeline-8                3328          2928          -12.02%
BenchmarkEncodingJSONMarshal_Simple_elastictranscoderListJobsByPipeline-8      120           120           +0.00%
BenchmarkRESTXMLBuild_Complex_cloudfrontCreateDistribution-8                   103376        102536        -0.81%
BenchmarkRESTXMLBuild_Simple_cloudfrontDeleteStreamingDistribution-8           9104          8272          -9.14%
BenchmarkEncodingXMLMarshal_Simple_cloudfrontDeleteStreamingDistribution-8     4656          4656          +0.00%
```